requirements.txt
duckdb==0.9.2
neo4j==5.14.1
qdrant-client==1.7.0
spacy==3.7.2
langchain==0.1.0
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
typing-extensions==4.9.0 